Increase wing angle range and adjust wing rotation axis
In the PaperPlaneScene.unity file, the local position of a game object was changed from {x: -15, y: 10, z: -40} to {x: -10, y: 10, z: -45}.

In the EvolutionManager.cs file, the range for the wingAngle property of the PlaneShape class was increased from 0f to 3f to 0f to 45f during initialization. Additionally, the clamping range for the wingAngle property during mutation was also increased from 0f to 3f to 0f to 45f.

In the PlaneShape.cs file, the application of the wingAngle to the local Euler angles of the left and right wings was modified. Instead of adjusting the x component of the local Euler angles, the y component is now adjusted.